+++ apps/cowswap-frontend/src/modules/twap/const.ts
@@ -1,7 +1,8 @@
 import { MessageDescriptor } from '@lingui/core'
 
 import { USDC } from '@cowprotocol/common-const'
-import { mapAddressToSupportedNetworks, mapSupportedNetworks, SupportedChainId } from '@cowprotocol/cow-sdk'
+import { isProdLike } from '@cowprotocol/common-utils'
+import { mapAddressToSupportedNetworks, mapChainEnum, SupportedChainId } from '@cowprotocol/cow-sdk'
 import { Currency, CurrencyAmount, Percent } from '@cowprotocol/currency'
 
 import { msg } from '@lingui/core/macro'
@@ -23,7 +24,7 @@ export const DEFAULT_NUM_OF_PARTS = 2
 
 export const DEFAULT_ORDER_DEADLINE: OrderDeadline = { label: msg`1 Hour`, value: ms`1 hour` }
 
-export const ORDER_DEADLINES: OrderDeadline[] = [
+const ORDER_DEADLINES_BASE: OrderDeadline[] = [
   DEFAULT_ORDER_DEADLINE,
   { label: msg`6 Hours`, value: ms`6 hour` },
   { label: msg`12 Hours`, value: ms`12 hour` },
@@ -31,6 +32,12 @@ export const ORDER_DEADLINES: OrderDeadline[] = [
   { label: msg`1 Week`, value: ms`1d` * 7 },
   { label: msg`1 Month`, value: ms`1d` * 30 },
 ]
+
+const DEV_ORDER_DEADLINE: OrderDeadline = { label: msg`4 Minutes`, value: ms`4min` }
+
+export const ORDER_DEADLINES: OrderDeadline[] = isProdLike
+  ? ORDER_DEADLINES_BASE
+  : [DEV_ORDER_DEADLINE, ...ORDER_DEADLINES_BASE]
 
 export const TWAP_ORDER_STRUCT = [
   {
@@ -62,14 +69,21 @@ export const TWAP_FINAL_STATUSES = [
   TwapOrderStatus.Cancelled,
 ]
 
-export const MINIMUM_PART_SELL_AMOUNT_FIAT: Record<SupportedChainId, CurrencyAmount<Currency>> = {
-  ...mapSupportedNetworks((chainId: SupportedChainId) => CurrencyAmount.fromRawAmount(USDC[chainId], 1e6)), // 1$ for most chains
-  [SupportedChainId.MAINNET]: CurrencyAmount.fromRawAmount(USDC[SupportedChainId.MAINNET], 1_000e6), // 1k for mainnet
-  [SupportedChainId.SEPOLIA]: CurrencyAmount.fromRawAmount(USDC[SupportedChainId.SEPOLIA], 10e18), // 10 for sepolia
-  [SupportedChainId.BNB]: CurrencyAmount.fromRawAmount(USDC[SupportedChainId.BNB], 1e18), // 1 for BNB, but it has 18 decimals!
-}
+export const MINIMUM_PART_SELL_AMOUNT_FIAT: Record<SupportedChainId, CurrencyAmount<Currency>> = isProdLike
+  ? {
+      ...mapChainEnum(SupportedChainId, (chainId: SupportedChainId) =>
+        CurrencyAmount.fromRawAmount(USDC[chainId], 1e6),
+      ), // 1$ for most chains
+      [SupportedChainId.MAINNET]: CurrencyAmount.fromRawAmount(USDC[SupportedChainId.MAINNET], 1_000e6), // $1000 for mainnet
+      [SupportedChainId.SEPOLIA]: CurrencyAmount.fromRawAmount(USDC[SupportedChainId.SEPOLIA], 10e18), // $10 for sepolia
+      [SupportedChainId.BNB]: CurrencyAmount.fromRawAmount(USDC[SupportedChainId.BNB], 1e18), // $1 for BNB, but it has 18 decimals!
+    }
+  : mapChainEnum(
+      SupportedChainId,
+      (chainId: SupportedChainId) => CurrencyAmount.fromRawAmount(USDC[chainId], 10 ** USDC[chainId].decimals), // $1 for all chains
+    )
 
-export const MINIMUM_PART_TIME = ms`5min` / 1000 // in seconds
+export const MINIMUM_PART_TIME = isProdLike ? ms`5min` / 1000 : ms`2min` / 1000 // in seconds
 export const MAX_PART_TIME = MAX_ORDER_DEADLINE / 1000 // in seconds
 
 export const DEFAULT_TWAP_EXECUTION_INFO: TwapOrderExecutionInfo = {