+++ libs/multicall/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [3.7.2](https://github.com/cowprotocol/cowswap/compare/multicall-v3.7.1...multicall-v3.7.2) (2026-09-15)
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @cowprotocol/cowswap-abis@4.2.0
+  - @cowprotocol/common-const@3.12.0
+  - @cowprotocol/wallet@3.13.2
+
 ## [3.7.1](https://github.com/cowprotocol/cowswap/compare/multicall-v3.7.0...multicall-v3.7.1) (2026-09-11)
 
 ### Patch Changes